refactor(types): split privacy.ts into entities/{bucket,friend-connection,friend-code,key-grant}.ts
